Only add validation to embedded redhen_contact element on user form if it is present.
This prevents fatal errors if the redhen element is removed from the
user register/edit form by a form alter (e.g. on the password reset
page).

To ensure this causes no issues with saving Drupal users, we had to
modify validation calls to ensure we’re not trying to access form
elements that don’t exist. Specifically - we were trying to access
$form[‘#form_id’] inside the redhen elements we were validating, which
didn’t exist because the element does not have the same array structure
as the entire from.