Change approach for disabling dependent fields on Contact settings form.
Previously we were disabling two fields if their dependent checkbox
wasn’t checked, but that makes for bad UX if the child field is
checked, then the parent field gets unchecked, and the end-user sees
the child field checked and disabled.

Instead, we’re now hiding the child fields when their parents are
unchecked and ensuring the correct field values are set in the submit
handler using && logic (e.g. B = TRUE if A == TRUE && B == TRUE).

#88